ecommerce_app/requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
certifi==2023.11.17
charset-normalizer==3.3.2
Django==4.2.7
django-bootstrap-v5==1.0.11
django-payments==2.0.0
django-phonenumber-field==7.2.0
idna==3.4
phonenumberslite==8.13.26
python-dotenv==1.0.0
requests==2.31.0
soupsieve==2.5
sqlparse==0.4.4
stripe==7.6.0
typing_extensions==4.8.0
urllib3==2.1.0
